Rename gafaelfawr.pydantic to gafaelfawr.types
This is a somewhat more accurate name for that internal module.
